pyproject.toml
[tool.poetry]
name = "team-flow-team-management-service"
version = "0.1.0"
description = ""
authors = ["dattar321 <dattar321@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-split-settings = "^1.3.1"
django-cors-headers = "^4.3.1"
psycopg = "^3.1.18"
djangorestframework-simplejwt = "^5.3.1"
psycopg-binary = "^3.1.19"
pika = "^1.3.2"
model-bakery = "^1.18.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"


[tool.poetry.group.test.dependencies]
coverage = "^7.5.1"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_unused_ignores = true
exclude = [
    '**\migrations\*.py',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

